add tests for sign operation
